refractoring binary search into algorithms folder
